refactor: adds barrel exports and updates scripts
