+++ package.json
@@ -7,12 +7,13 @@
     "yarn": ">=1.22.19"
   },
   "scripts": {
-    "build": "NODE_ENV=production tsc",
-    "dev": "NODE_ENV=development tsx watch src/index.ts",
-    "db:generate:dev": "NODE_ENV=development drizzle-kit generate",
-    "db:push:dev": "NODE_ENV=development drizzle-kit push",
-    "db:studio:dev": "NODE_ENV=development drizzle-kit studio",
-    "db:seed:dev": "NODE_ENV=development tsx src/scripts/seed.ts"
+    "build": "tsc",
+    "dev": "tsx watch src/index.ts",
+    "db:generate:dev": "drizzle-kit generate",
+    "db:push:dev": "drizzle-kit push",
+    "db:studio:dev": "drizzle-kit studio",
+    "db:seed:dev": "tsx src/scripts/seed.ts",
+    "start": "node dist/src/index.js"
   },
   "dependencies": {
     "@hono/node-server": "^1.11.3",
