+++ src/config/env.ts
@@ -1,5 +1,3 @@
-import path from "path";
-import fs from "fs";
 import { z } from "zod";
 import { config } from "dotenv";
 
@@ -11,23 +9,23 @@ const environmentSchema = z.object({
   JWT_SECRET: z.string(),
 });
 
-const getEnvironmentPath = () => {
-  return path.resolve(new URL(import.meta.url).pathname, `../../../.env.${process.env.NODE_ENV}`);
-};
+// const getEnvironmentPath = () => {
+//   return path.resolve(new URL(import.meta.url).pathname, `../../../.env.${process.env.NODE_ENV}`);
+// };
 
-const loadEnvironmentVariables = () => {
-  const environmentPath = getEnvironmentPath();
+// const loadEnvironmentVariables = () => {
+//   const environmentPath = getEnvironmentPath();
 
-  if (!fs.existsSync(environmentPath)) {
-    console.warn(`Environment variables file for the current environment is not found.`);
-    return;
-  }
+//   if (!fs.existsSync(environmentPath)) {
+//     console.warn(`Environment variables file for the current environment is not found.`);
+//     return;
+//   }
 
-  config({ path: environmentPath });
-};
+//   config({ path: environmentPath });
+// };
 
 const validateEnvironment = () => {
-  loadEnvironmentVariables();
+  // loadEnvironmentVariables();
 
   const parsedEnvironment = environmentSchema.safeParse(process.env);
